refactor(review): 별점 타입을 Short에서 BigDecimal로 변경
- 별점 범위를 1~5에서 1.0~5.0으로 확장
- Review 엔티티, DTO, Service 전반에 걸쳐 관련 필드 및 로직 수정
- 입력 검증에 소수점 한 자리까지 허용하는 DecimalMin, DecimalMax, Digits 제약 추가

+++ src/main/java/com/sodosiro/domain/review/controller/dto/response/ReviewResponse.java
@@ -6,6 +6,7 @@ import com.sodosiro.domain.review.constants.ReviewVisitType;
 import com.sodosiro.domain.travel.entity.TouristSpot;
 import com.sodosiro.domain.user.entity.User;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,7 +14,7 @@ public record ReviewResponse(
         Long reviewId,
         AuthorInfo author,
         SpotSummary spot,
-        Short rating,
+        BigDecimal rating,
         String body,
         List<ReviewImageResponse> images,
         LocalDateTime createdAt,
